% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegLG.R
\name{regl}
\alias{regl}
\title{Regresión Logistica}
\usage{
regl(xp, x, y, clase_nombre)
}
\arguments{
\item{xp}{(vector) muestra de datos de la variable predictora}

\item{x}{(vector) variable predictora}

\item{y}{(vector) muestra de la variable respuesta}

\item{clase_nombre}{nombre de la variable dependiente}
}
\value{
un data frame con las probabilidades requeridas
}
\description{
Realiza una regresión logistica
}
\examples{
\dontrun{
#Directorio de trabajo
ruta <- "/Users/jg289/Downloads/dataset1.csv"

#primer ejemplo
datos <- read.csv(ruta)
#Cargo la libreria
library(RegresionLog)
regl(substr(df$evi,1,100),df$evi,substr(df$presencia,1,100),"presencia")
}
}
